backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
